Prevent spurious branches
In case a CSR (or any system, actually) instruction was
directly executed after a taken branch, the csr would also
incorrectly branch to somewhere.
Because BCU was still enabled and thus x_taken would still be
possibly set to 1 which would therefore set m_taken to 1 also.

To prevent that, just bare in mind to set d_bcu_op in ALL if branch.

With this commit, csr unit tests are now passing.