make: use 'xxd -p' instead of parsing output
The default output of xxd is meant for humans, but the expected format
for .hex inclusion is much simpler.
The output of xxd was thus parsed to strip all human-related formatting,
but it turns out xxd already has an option to do that.

This patch makes use of that option in order to simplify the top
Makefile.